Rename UserCreateForm to SignUpForm
UserCreateForm was supposed to server the dual purpose of sign up and edit account. However, since the control logic for Editing Account is very different from that of a Sign Up Form
I decided to separate the two. The differences lie mainly in that a Create Account would require a POST request, but an edit account requires a PATCH request.